added sanity check if value references and variable names are unique, some fixes based on coverity scan, some more cosmetic changes

+++ import/integrators/src/IntegratorStepper.cpp
@@ -674,7 +674,7 @@ public:
 
 		//CVodeSetErrFile( cvode_mem, NULL ); // uncomment to suppress error messages
 
-		CVodeSetMaxNumSteps( cvode_mem_, 1.0e5 );
+		CVodeSetMaxNumSteps( cvode_mem_, static_cast<long>( 1.0e5 ) );
 	}
 
 	~SundialsStepper()
